fix bonus protection but leaks hh

+++ src/bonus/here_doc_bonus.c
@@ -6,7 +6,7 @@
 /*   By: waraissi <waraissi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/14 16:44:11 by waraissi          #+#    #+#             */
-/*   Updated: 2023/01/21 15:19:49 by waraissi         ###   ########.fr       */
+/*   Updated: 2023/01/22 19:22:26 by waraissi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,18 +17,19 @@ void	execution_sec(t_heredoc *vars, int *fd, char **cmd1)
 	int	id;
 
 	id = fork();
+	if (id == -1)
+		exit(1);
 	if (id == 0)
 	{
-		close(fd[1]);
-		dup2(fd[0], 0);
-		close(fd[0]);
-		dup2(vars->outfile, 1);
+		if (close(fd[1]) == -1 || dup2(fd[0], 0) == -1 || close(fd[0]) == -1
+			|| dup2(vars->outfile, 1) == -1)
+			exit(1);
 		execute_child(vars, cmd1);
 	}
 	else
 	{
-		close(fd[0]);
-		close(fd[1]);
+		if (close(fd[0]) == -1 || close(fd[1]) == -1)
+			exit(1);
 		waitpid(id, 0, 0);
 	}
 }
@@ -38,20 +39,22 @@ void	execution(t_heredoc *vars, int tmp, char **cmd, char **cmd1)
 	int	fd[2];
 	int	id;
 
-	pipe(fd);
+	if (pipe(fd) == -1)
+		exit(1);
 	id = fork();
+	if (id == -1)
+		exit(1);
 	if (id == 0)
 	{
-		close(fd[0]);
-		dup2(tmp, 0);
-		close(tmp);
-		dup2(fd[1], 1);
-		close(fd[1]);
+		if (close(fd[0]) == -1 || dup2(tmp, 0) == -1 || close(tmp) == -1
+			|| dup2(fd[1], 1) == -1 || close(fd[1]) == -1)
+			exit(1);
 		execute_child(vars, cmd);
 	}
 	else
 	{
-		close(tmp);
+		if (close(tmp))
+			exit(1);
 		execution_sec(vars, fd, cmd1);
 	}
 }
@@ -59,27 +62,38 @@ void	execution(t_heredoc *vars, int tmp, char **cmd, char **cmd1)
 void	execute_heredoc(t_heredoc *vars, char **cmd, char **cmd1)
 {
 	int		tmp;
-	char	*line;
-	char	*limiter;
 
 	tmp = open(".heredoc.txt", O_RDWR | O_CREAT | O_TRUNC, 0644);
 	if (!tmp)
 		return ;
 	while (1)
 	{
-		write(1, "> ", 2);
-		line = get_next_line(0);
-		limiter = ft_strjoin(vars->h_d, "\n");
-		if (line == NULL || ft_strcmp(line, limiter) == 0)
+		if (write(1, "> ", 2) == -1)
+			exit(1);
+		vars->line = get_next_line(0);
+		vars->limiter = ft_strjoin(vars->h_d, "\n");
+		if (vars->line == NULL || ft_strcmp(vars->line, vars->limiter) == 0)
 			break ;
-		write(tmp, line, ft_strlen(line));
-		free(line);
-		free(limiter);
+		if(write(tmp, vars->line, ft_strlen(vars->line)) == -1)
+			exit(1);
+		free(vars->line);
+		free(vars->limiter);
 	}
-	close(tmp);
+	if (close(tmp) == -1)
+		exit(1);
 	tmp = open(".heredoc.txt", O_RDONLY);
+	if (!tmp)
+		return ;
 	execution(vars, tmp, cmd, cmd1);
 	unlink(".heredoc.txt");
+}
+
+void	initialze(t_heredoc *vars, int ac, char **av, char **envp)
+{
+	vars->outfile = ac - 1;
+	vars->l_a = av[ac - 1];
+	vars->h_d = av[2];
+	vars->envp = envp;
 }
 
 void	here_doc(int ac, char **av, char **envp)
@@ -89,10 +103,7 @@ void	here_doc(int ac, char **av, char **envp)
 	char		**cmd1;
 	char		*path;
 
-	vars.outfile = ac - 1;
-	vars.l_a = av[ac - 1];
-	vars.h_d = av[2];
-	vars.envp = envp;
+	initialze(&vars, ac, av, envp);
 	if (ac == 6 && !ft_strncmp(av[1], "here_doc", sizeof(av[1])))
 	{
 		check_last_arg(&vars);
